+++ stories/patterns/media/templates/image-with-caption.js
@@ -1,0 +1,18 @@
+export default `
+<figure>
+    <VsImg 
+        src="./fixtures/image-with-caption/images/blair-castle.webp"
+        fluid
+    />
+    <figcaption>
+        <VsMediaCaption>
+            <template v-slot:caption>
+                Blair Castle, Blair Atholl, Perthshire
+            </template>
+            <template v-slot:credit>
+                © VisitScotland / Paul Tomkins
+            </template>
+        </VsMediaCaption>
+    </figcaption>
+</figure>
+`;